diagram_generator/requirements.txt
railroad-diagrams
cairosvg
